Remove SERIALIZATION_SPLIT_MEMBER
SERIALIZATION_SPLIT_MEMBER generates a
`template<typename Archive> sample::serialize(Archive, uint)` function that
tries to detect the type of archive it is called with and then calls either
`save()` or `load()`. As the `Archive` is either `eos::portable_iarchive` *or*
`eos::portable_oarchive`, we can just define `serialize()` as overloaded
function and avoid the overhead altogether.