Introduce soname to indicate ABI version of shared library
Having a "soname" that stays constant allows compiled
executables that link against liblsl to continue to work
whithout recompilation when liblsl is updated as long as
the soname of liblsl stays the same.

Each time the ABI of liblsl changes in future, the setting
LSL_ABI_VERSION in CMakeLists.txt should be incremented by
one to change the soname.  Closes #140.